assignr: Create Homework Assignments and Solutions using 'RMarkdown'

Writing homework assignments for students in the age of 'RMarkdown' necessitates the creation of two separate documents – assign.Rmd and soln.Rmd. The goal of assignr is to create one document main.Rmd that can be broken apart into the above two documents. Thus, there is no longer a need to copy and paste between the assign and the soln documents as all of the contents are together in one file.

See also

Author

Maintainer: James Balamuta balamut2@illinois.edu [copyright holder]

Authors:

David Dalpiaz dalpiaz2@illinois.edu [copyright holder]